src/requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
ecdsa==0.17.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
passlib==1.7.4
protobuf==3.19.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-dotenv==0.19.2
python-jose==3.3.0
PyYAML==6.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing-extensions==4.0.0
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
zipp==3.6.0
